pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0.post1"}
sqlalchemy = "^2.0.25"
psycopg2 = "^2.9.9"
alembic = "^1.13.1"
pydantic-extra-types = "^2.5.0"
phonenumbers = "^8.13.29"
pydantic = {extras = ["email"], version = "^2.6.0"}
faker = "^22.6.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
libgravatar = "^1.0.4"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.7"
bcrypt = "4.0.1"
fastapi-mail = "^1.4.1"
fastapi-limiter = "^0.1.6"
redis = "^5.0.2"
cloudinary = "^1.39.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
